Try to refresh an expired identity token
When the identity token expires, so far the user needed to do an
(interactive) login with their IDP. This is inconvenient and causes
issues with background requests in Neos, making them fail.

This change stores a refresh token in the user session, so as long as
the user does not log out from Neos, an expired identity token will be
refreshed if possible.

Fixes #15